+++ CHANGELOG.md
@@ -18,6 +18,8 @@
 
 ### Updated
 
+- dependencies
+
 ## [0.13.0] - 2025-06-19
 
 ### Fixed